Optimization of the CoreData fetch

+++ WaterLogging/WaterLogging/Models/WaterLogProgress.swift
@@ -11,5 +11,5 @@ import Foundation
 /// A value to store today's water consumption progress towards a goal
 struct WaterLogProgress {
     /// Represents an amount in millileters (mL) that the user has logged in water intake
-    let amount: Double
+    let amount: Int
 }
